eval_encoder/OpenTAD/requirements.txt
###### Requirements without Version Specifiers ######
mmengine
wandb
scipy
einops
pandas
tqdm
ninja
imgaug
pytorchvideo

###### Requirements with Version Specifiers ######
numpy==1.23.5
gdown==5.1.0

###### Custom Package Requirements ######
# opentad/models/utils/post_processing/nms 
# opentad/models/roi_heads/roi_extractors/align1d
# opentad/models/roi_heads/roi_extractors/boundary_pooling